refactor(main): extract shared step components (SectionLabel, ResultKbd, ResultAnnouncement)
Deduplicate identical patterns across vocabulary-step, multiple-choice-step, and sort-order-step into reusable components for consistency as new step types are added.
