+++ apps/main/src/components/activity-player/section-label.tsx
@@ -1,0 +1,5 @@
+export function SectionLabel({ children }: { children: React.ReactNode }) {
+  return (
+    <p className="text-muted-foreground text-xs font-medium tracking-wide uppercase">{children}</p>
+  );
+}
